Made changes to the theme, using use-package, and general QoL changes.
Changed the theme because dracula makes more changes to more faces,
and keeps things looking good. It also mixed well with the
header-based theme I've got going, after a bit of massaging. A huge
change is the use of Iosevka (Sparkle) as the font family of
choice. It's got the right mix of condensed but still beautiful to
read.

Use-package is far more useful than I gave it credit for, and the
change was worthwhile. It's made it much easier to defer package
loading and keeps hooks tidy.

LSP mode has made proper Powershell Intellisense possible, and opens
up the ability to add more language support later.