Commented out unicode-fonts, pointed things to MELPA, deferred thing
Unicode Fonts solves a problem I don't feel I have. I'll need to play
around without it before I'm comfortable removing it.

Pointed company-mode and org-mode away from their master branches in
their dev repos in favor of whatever's on MELPA. It's easier to make
sure everything is pulling from the right place if it all says
`:straight t` in it.

Deferred loading recentf until it's needed. Checking to see if this
makes sense to do or not.